Backend/requirements.txt
json-rpc==1.13.0
pytest==7.4.0
python-dotenv==1.0.0

